manager -> manager.ptr in ccall to fix null ptr issue (#8)
## Summary of Changes

In the current codebase, we're incorrectly passing `manager::Manager` into several of the `@ccall` which actually expects `manager.ptr::ManagerPtr`. On MacOS, this didn't cause an error because `manager.ptr` happens to locate at the beginning of the struct, so referencing the outer struct yields the correct ptr. However, this behavior is very compiler-dependent, and on Ubuntu 24.04 LTS, these lines cause the downstream application to crash due to null ptr dereference.

This PR fixes the issue by replacing the two places where `manager::Manager` is used with `manager.ptr::ManagerPtr`